Finished adding wl_region
As of now it is not used anywhere, it pretty much just implements
the data structs for holding a list of rectangles. It needs to be
incorporated to the damage/input code.

+++ compositor/src/category5/utils/region.rs
@@ -7,8 +7,8 @@
 // This an offset of unspecified units from
 // some basis. Basically a cartesian point.
 pub struct Offset2D {
-    pub x: f64,
-    pub y: f64,
+    pub x: i32,
+    pub y: i32,
 }
 
 // A rectangular region
